+++ CHANGELOG.md
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 	- subprocess errors include exit code
 	- logs unexpected stdout instead of triggering E_USER_NOTICE (via optional logger)
 	- logs unexpected stderr instead of throwing an exception (via optional logger)
+- `JobInfo`
+	- `getExtendedExpression()` includes seconds only if seconds are used
 
 ### Fixed